@@ -3,7 +3,8 @@
   "private": true,
   "version": "0.0.1",
   "engines": {
-    "node": ">=16"
+    "node": ">=16",
+    "pnpm": ">=6"
   },
   "scripts": {
     "dev": "vite",
@@ -11,12 +12,12 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "devDependencies": {
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@vitejs/plugin-react": "^1.0.7",
     "typescript": "^4.5.4",
     "vite": "^2.9.0"
