@@ -1,11 +1,14 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
+import * as React from 'react'
+import * as ReactDOMClient from 'react-dom/client'
 import './index.css'
 import App from './App'
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-)
+function Root() {
+  return (
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+  )
+}
+const root = ReactDOMClient.createRoot(document.getElementById('root')!)
+root.render(<Root />)
